requirements.txt
numpy==1.24.3
opencv_contrib_python==4.9.0.80
opencv_python==4.7.0.72
Pillow==8.1.0
PyQt5==5.15.10
scikit_image==0.19.2
setuptools==63.4.1
torchvision==0.15.1
wand==0.6.11